Lab2_REST_with_in_memory/src/goal_task_service/requirements.txt
fastapi
uvicorn
pydantic
pyjwt
passlib[bcrypt]
bcrypt
python-jose[cryptography]
python-dotenv